Convert synthetic data to parquet with LAC-specific formatting (#713)
* add LAC-specific formatting

* update synthetic data generation

* generate the updated synthetic data

* removed outdated files

* fix casing for FHIR
